카프카 도입 (컨슈머 프로젝트 추가) - 컨슈머 프로젝트 추가 - 카프카 커넥터 관련 설정 삭제

+++ bidder/src/main/java/com/example/mybidder/bidding/service/BidService.java
@@ -40,7 +40,7 @@ public class BidService {
                                                 new KafkaBidLog(
                                                         bidRequest.requestId(),
                                                         campaign.id(),
-                                                        MicroConverter.convertMicroToCpm(campaign.targetCpmMicro())
+                                                        MicroConverter.convertMicroToCpm(campaign.targetCpmMicro()).doubleValue()
                                                 ));
                                         return Mono.just(new BidResponse(
                                                 bidRequest.requestId(),
